6.4.2 ProxyFactoryBean 적용
