--- source/yuuuunmi/src/main/java/springbook/user/service/TxProxyFactoryBean.java
@@ -1,52 +1,0 @@
-package springbook.user.service;
-
-import org.springframework.beans.factory.FactoryBean;
-import org.springframework.transaction.PlatformTransactionManager;
-
-import java.lang.reflect.Proxy;
-
-public class TxProxyFactoryBean implements FactoryBean<Object> {
-    private Object target;
-    private PlatformTransactionManager transactionManager;
-    private String pattern;
-    private Class<?> serviceInferface;
-
-    public void setTarget(Object target) {
-        this.target = target;
-    }
-
-    public void setTransactionManager(PlatformTransactionManager transactionManager) {
-        this.transactionManager = transactionManager;
-    }
-
-    public void setPattern(String pattern) {
-        this.pattern = pattern;
-    }
-
-    public void setServiceInferface(Class<?> serviceInferface) {
-        this.serviceInferface = serviceInferface;
-    }
-
-    @Override
-    public Object getObject() throws Exception {
-        TransactionHandler txHandler = new TransactionHandler();
-        txHandler.setTarget(target);
-        txHandler.setTransactionManager(transactionManager);
-        txHandler.setPattern(pattern);
-
-        return Proxy.newProxyInstance(
-                getClass().getClassLoader(),
-                new Class[]{serviceInferface},
-                txHandler);
-    }
-
-    @Override
-    public Class<?> getObjectType() {
-        return serviceInferface;
-    }
-
-    @Override
-    public boolean isSingleton() {
-        return false;
-    }
-}
